[Turbopack] improve task optimization (#74837)
### What?

improves the heuristic for task optimization.

It minimizes the produce of `(uppers - 1) * follower` and chooses a bigger aggregation number based on that.

This allows for many uppers when there are no followers. That is often the case when a task is shared between multiple pages. Or for tasks are are near the leafs of the graph.