fix(turbopack-swc-utils): Don't panic if issues aren't emitted, use `#[must_use]` annotation (#74906)
This is creating a ton of logspam for me when running tests. I can get dozens of these panics in a row.

It appears that there are functions that can return an `anyhow::Error` between collecting issues and emitting them. In those cases, it's not practical to ensure the issues are always emitted.

Instead, just settle for a much weaker compile-time check with `#[must_use]` that at least ensures something happens with `IssueCollector`.