Fix source mapping of generated cache wrapper calls
In the SWC transform we're now using the span of the original `"use
cache"` function for the generated cache wrapper call, e.g. so that the
captured timeout error stack can be source mapped correctly.

This also fixes the wrongly placed
`/*#__TURBOPACK_DISABLE_EXPORT_MERGING__*/` marker for the generated
cache wrapper statements.