[Segment Cache] Evict client cache on revalidate (#74874)
This implements evicting the client cache when a Server Action calls
revalidatePath or revalidateTag.

Similar to the old prefetching implementation, it works by clearing the
entire client cache, as opposed to only the affected path or tags. There
are more changes needed on the server before we can support granular
cache eviction. This just gets us to parity with the status quo.