Add `next-custom-transforms` tests to verify source maps
This allows us to quickly verify that the spans that are defined in the
SWC transforms yield correct source maps. The `output.js` and
`output.js.map` files can be used to visualize the source mapping, e.g.
at https://evanw.github.io/source-map-visualization/.

For now I'm adding a single fixture for `'use cache'` with this PR.