perf(turbo-tasks): Use the type information we already have for all ResolvedVc casts, expose synchronous versions of functions (#75055)
We already have type information for `ResolvedVc`, so we can avoid going through the codepath in `Vc`, which can fail and requires a cell read.

I did this for sidecasts in https://github.com/vercel/next.js/pull/74844 to unblock synchronous graph traversals. This does the same for the rest of the functions.

I expect any perf impact to be pretty small (casting isn't a particularly hot codepath), but it's also an easy change to make, and (once callsites are migrated) will lead to a nicer API.

Right now I'm continuing to expose the async fallible versions of these functions. I'll port the callsites (looks like there's 92 of them) to the synchronous infallible version in a later PR.