fix: be more defensive in useMergedRef (#75088)
- Fixes a crash when a Link unmounts (under some VERY specific
conditions)
- Reverts #75012, which implemented a workaround for the crash

### Short explanation

The crash happens because we use a cleanup-returning ref in (the app-dir
version of) Link, and `<Link legacyBehavior>` "leaks" that ref into user
code. If user code does ref-merging incorrectly, i.e. it treats all
callback refs as the old style that gets called with `null`, then our
ref can get called with `null` even though it shouldn't, and crashes.

Technically this crash is not a Next.js bug. But I think a lot user code
uses ref-merging libraries that haven't been updated for 19, and can
crash because of this. So i think the safest course of action is to be
defensive, and change `useMergedRef` to always convert cleanup refs to
normal function refs, so that userspace can't mess up our refs.

### Full explanation

- `observeLinkVisibilityOnMount` inside our Link is a cleanup-returning
ref
- i.e. the new thing added in react 19, `(el: Element) => () =>
void` instead of the old `(el: Element | null) => void`
- user code does  `<Link legacyBehavior><CustomComponent /></Link>`
- `legacyBehavior` means that props get spread onto the child,
including `ref`, so we'll pass the ref created inside Link
to  `CustomComponent`
- AND no `ref` is passed on the `<CustomComponent>`  itself
- with `legacyBehavior`, that'd get extracted and merged with
`observeLinkVisibilityOnMount`
- if there's no user ref, the old implementation of `useMergedRef` would
return `observeLinkVisibilityOnMount` unchanged, thus
passing `CustomComponent` a cleanup ref
- `CustomComponent` uses a buggy ref merging library, i.e. one that
wasn't updated to handle cleanup refs.
- **so `observeLinkVisibilityOnMount` gets treated as an old-style
callback ref, i.e. it gets called with`null`. and stuff explodes,
because it doesn't expect that**

The test in `test/production/next-link-legacybehavior-ref-merging`
reproduces the crash. You can check out the commit that adds it to see
it happen -- I put it before the fix.