[Segment Cache] Remove segment access tokens (#75157)
Previously I implemented a lightweight "access token" that the client
must pass when prefetching a segment. The tokens were a hash of the
parent segment path; the idea was to eventually add a secure salt to
make the segment path's unguessable. Not for security, really, just as
an extra precaution against Denial of Wallet attacks.

However, since we're not doing that yet, the access token as currently
implemented doesn't do anything in practice besides add extra bytes to
the requests.

The more immediate issue, though, is that because access tokens can't be
inferred from the original URL, we need to add something to the
prerender manifest so we can rewrite incoming requests to the correct
path, which is a bit tricky in the case of parameterized URLs because
the parent param values are part of the hash. To add this feature back
in the future, we'll need to account for this somehow.

Since this is not an urgent feature, I'm removing it for now.