[Segment Cache] Minimize special root key handling (#75159)
The root segment key is a special case: it's an empty string, so that
it's not part of any of the child paths, but the response for the root
segment itself is written to the filesystem as `_index`. I only did this
to make debugging easier; it would be fine to remove the special case
such that every segment path begins with `_index`. It's just a bit
ugly/redundant.

But my main concern is that the special case handling does not leak
outside of Next.js, e.g. into the Vercel builder. We should aim to keep
the protocol as simple as possible. (However, I think it'd be fine if it
leaked into the prerender manfiest, since that's generated by Next.js
itself.)

For now, this moves the special case handling of the root key to as few
places as possible: 1) when generating the prefetch responses, and 2)
when the client issues a request. In all other places, the segment keys
should be treated as arbitrary strings/paths.