[Segment Cache] Predictable fallback param encoding (#75166)
When a segment includes a fallback parameter, during prerendering, a
random placeholder is used to generate the response. We then use this
response to generate a key path for each segment. We need this path to
be predictable so the server can create a rewrite for it; it can't
include random placeholders.

So, this updates the segment generation to replace the placeholder
values with a "template" string, e.g. `[paramName]`.