test: skip dev/deploy testing for `test/e2e/app-dir/segment-cache/revalidation/segment-cache-revalidation.test.ts` (#75249)
### Why?

Since this test creates a mock server to fetch before creating a Next.js
instance, it will fail on the deployment test due to a mismatch in the
location of the server and the instance. The test did try to skip when
deployed but it was after the `createNext` logic, which triggered the
deployment before skipping.

### How?

Move the skip before any `beforeAll` logic like creating an instance.

---------

Co-authored-by: Zack Tanner <1939140+ztanner@users.noreply.github.com>