Update not-found.mdx
This PR updates the page example to reflect the changes introduced in Next.js 15 regarding access to route parameters. 

Specifically, the example will be modified to include the async keyword in the page function definition. This change is necessary because, as of Next.js 15, params is a Promise that needs to be resolved asynchronously. 

The update will ensure that the documentation aligns with Next.js v15 code.