tweak static generation retry wait time (#75381)
Currently when retrying workers we wait somewhere between 0-500ms. This
is to add a variable amount of jitter but the timeout is still fairly
short. This tweaks it to ensure we're somewhere between 500-2000ms and
adds a jitter value up to 30% of the base delay value to ensure large
amounts of SSG failures don't cause a thundering herd problem on retry.