[Turbopack] disable lifo slot to make all tasks stealable (#75472)
### What?

* pass rustflags to windows too
* disable lifo slot to make all tasks stealable

### Why?

Since we do a lot of CPU bound work in tokio tasks this can lead to future polling taking several seconds. With the lifo slot, one task is not stealable and need to wait for these several seconds to be executed. This delays that task and causes reduced parallelization of tasks.