[DevOverlay] fix total error count (#75500)
`readyErrors` seems to be the value that we're using to actually display
the errors in the overlay, while `state.errors` seems to often contain
duplicates. This causes a drift in the number of errors reported in the
overlay vs the indicator. We should use the same source of truth for
both.