Disable `rewrite-headers` tests that are dependent on Vercel site config (#75647)
The two scenarios that test rewrite headers for vercel.com are dependent
on how the Vercel site has currently configured its rewrites, and thus
are prone to be flaky. Until we've figured out what Next.js logic this
is supposed to test, and how we can decouple it from this external
source, we are disabling them to unblock other PRs.

An attempt to fix the expectation was done yesterday in #75602, but in
the meantime the returned `x-nextjs-rewritten-path` header has changed
again
(`/home/eyJhbGciOiJIUzI1NiJ9.AP0.7cQg-Q2tj1BAyHFfT32HLI3JRdEfYCiANjbtqRiDfLs`).