feat: added support for prefetch segments when deployed (#75202)
Designed to enable the client prefetch segments feature when deployed on
Vercel, this adds additional fields and data to the manifests to allow
the CLI to create the rewrite rules needed to serve the prefetch
segments. The fallback routes also now emit their RSC data to disk even
though only their filenames are used.

We emit the routes for the prefetches using backreferences. For example,
for the route:

```
/blog/[...slug].segments/$c$slug$[...slug]/__PAGE__.segment.rsc
```

We create the following Regular Expression:

```ts
/^\/blog\/(?<nxtPslug>.+?)\.segments\/\$c\$slug\$\k<nxtPslug>\/__PAGE__\.segment\.rsc$/
```

Which enforces that the named capture group `nxtPslug` is the same for
both the `[...slug].segments` and the `$c$slug$[...slug]`:

```
Matching:
/blog/my-post.segments/$c$slug$my-post/__PAGE__.segment.rsc
/blog/hello-world.segments/$c$slug$hello-world/__PAGE__.segment.rsc
/blog/2024-01.segments/$c$slug$2024-01/__PAGE__.segment.rsc

Not Matching
/blog/post-1.segments/$c$slug$post-2/__PAGE__.segment.rsc
```

This also fixed a bug where the client segment prefetch keys included
double `//`.

<!-- Thanks for opening a PR! Your contribution is much appreciated.
To make sure your PR is handled as smoothly as possible we request that
you follow the checklist sections below.
Choose the right checklist for the change(s) that you're making:

## For Contributors

### Improving Documentation

- Run `pnpm prettier-fix` to fix formatting issues before opening the
PR.
- Read the Docs Contribution Guide to ensure your contribution follows
the docs guidelines:
https://nextjs.org/docs/community/contribution-guide

### Adding or Updating Examples

- The "examples guidelines" are followed from our contributing doc
https://github.com/vercel/next.js/blob/canary/contributing/examples/adding-examples.md
- Make sure the linting passes by running `pnpm build && pnpm lint`. See
https://github.com/vercel/next.js/blob/canary/contributing/repository/linting.md

### Fixing a bug

- Related issues linked using `fixes #number`
- Tests added. See:
https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs
- Errors have a helpful link attached, see
https://github.com/vercel/next.js/blob/canary/contributing.md

### Adding a feature

- Implements an existing feature request or RFC. Make sure the feature
request has been accepted for implementation before opening a PR. (A
discussion must be opened, see
https://github.com/vercel/next.js/discussions/new?category=ideas)
- Related issues/discussions are linked using `fixes #number`
- e2e tests added
(https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs)
- Documentation added
- Telemetry added. In case of a feature if it's used or not.
- Errors have a helpful link attached, see
https://github.com/vercel/next.js/blob/canary/contributing.md


## For Maintainers

- Minimal description (aim for explaining to someone not on the team to
understand the PR)
- When linking to a Slack thread, you might want to share details of the
conclusion
- Link both the Linear (Fixes NEXT-xxx) and the GitHub issues
- Add review comments if necessary to explain to the reviewer the logic
behind a change

### What?

### Why?

### How?

Closes NEXT-
Fixes #

-->

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>