Add test to verify which pages are prerendered in `use-cache` test suite (#75681)
This allows us to catch regressions and monitor progress (e.g. for NAR-85).

In addition, we're consolidating all pages in the test app to use a single root layout without a suspense boundary. Before we switched the app from using the `dynamicIO` flag to the `useCache` flag, some pages needed a suspense boundary to not fail the build.