`rootParams()` is now a plain Proimse (not exotic) (#75717)
When we made Request data APIs async we made them exotic to make the
transition to the Promise API easier to adopt because lots of existing
code probably accessed these values synchronously. `rootParams` is a new
Request Data API however and it has no existing usage that is async so
we do not need to support exotic access to these values. This change
removes the exotic construction in favor of a simple resolved Promise.