Respect `no-store` request header for `"use cache"` entries in dev mode
In dev mode, when doing a hard refresh or using dev-tools with "disable
cache" checked, the browser requests the page with a `cache-control:
no-store` header. When this is the case, we ignore existing cache
entries for `"use cache"` functions, and generate new ones instead. This
is now the same behavior as for `fetch` caches.

closes NAR-58