Dedupe cached data from `"use cache"` functions in RSC payload
By leveraging `React.cache` internally, we can ensure that multiple
invocations of the same `"use cache"` function return the same object
reference. This also allows React to dedupe cached data that's included
in the RSC payload, e.g. because it's passed to client components, which
will reduce the response sizes for the HTML and RSC requests.

closes NAR-53