Allow `"use cache"` with `output: 'export'`
Previously, when using `"use cache"` together with `output: 'export'`,
we were failing the build with the error `Server Actions are not
supported with static export.`

Now, instead of failing the build with any entry in the server reference
manifest, we only fail the build if the manifest includes a server
action, and allow the presence of server references for `"use cache"`
functions. We can differentiate those by looking at the server reference
information byte at the start of the server reference ID.

Notably, we're currently not preventing a user from using a `"use
cache"` function on the client. This would result in an error at
runtime.

closes NAR-77