[DevOverlay]: fix missing Error context for build-time errors (#75932)
`Error` should not be an optional property of the toolbar or error
overlay layout, since that information is needed for parsing the message
to determine if a docs link should be shown or if a stack trace can be
copied.

This marks those properties as non-optional so that TypeScript will warn
us in spots that were missing it. I updated the missing spots to
construct an error where one was not available.

Build time errors are also not meant to be dismissible. This forces the
error open until the build error is remedied.

Closes NDX-819